requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0

# 3D geometry processing
open3d>=0.16.0
trimesh>=3.15.0

# Computer vision and ML
opencv-python>=4.6.0
scikit-learn>=1.1.0
scikit-image>=0.19.0
Pillow>=9.2.0

# MMPose and related
mmengine>=0.8.0
mmcv>=2.0.1
mmdet>=3.1.0
mmpose>=1.1.0
xtcocotools>=1.12.0

# Data handling
pandas>=1.4.0
pyyaml>=6.0
tqdm>=4.64.0
h5py>=3.7.0

# Visualization (optional)
matplotlib>=3.5.0
plotly>=5.10.0

# System utilities
pathlib2>=2.3.0
argparse-color-formatter>=1.2.2
colorama>=0.4.5

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=5.0.0
